Removed some repetition, the Client and Server objects have been replaced with a single endpoint object. Also, a new test project has been written that replaces Test.Server and Test.Client, it also implements a WPF UI.

+++ Source/netSharp.Server/Connectivity/Session.cs
@@ -31,13 +31,12 @@ namespace netSharp.Server.Connectivity
         public bool IsDisposed { get; set; }
         public string RemoteEndpointGuid { get; set; }
         public string LocalEndpointGuid { get; set; }
-        public int IdleTime { get; set; }
-        public int MaxIdleTime { get; set; }
+        public double IdleTime { get; set; }
+        public double MaxIdleTime { get; set; }
 
         public void Dispose()
         {
             ctSource.Cancel();
-
             IsDisposed = true;
         }
